% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mrio_labels.R
\name{add_mrio_labels}
\alias{add_mrio_labels}
\title{Add MRIO labels}
\usage{
add_mrio_labels(.data, .col, from, to, replace = FALSE)
}
\arguments{
\item{.data}{A data frame.}

\item{.col}{Reference column.}

\item{from}{Field type of reference column.}

\item{to}{Target field type.}

\item{replace}{If TRUE, target replaces reference column. Default = FALSE
adds target as a new column next to the reference column.}
}
\value{
An object of the same type as `.data`.
}
\description{
Convert between various labels for countries and sectors in the
Multiregional Input-Output (MRIO) tables.
}
\details{
See ?dict_countries and ?dict_sectors for valid field types to
  convert between.
}
\examples{
df <- data.frame(
  t = c(2022, 2022, 2022),
  s = c(11, 25, 46),
  value = c(100, 90, 120)
)
df |> add_mrio_labels(s, from = "mrio_ind", to = "mrio_name")

# A named vector renames the new column.
df |> add_mrio_labels(s, from = "mrio_ind", to = c("country" = "mrio_name"))

}
