% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_eora}
\alias{cf_eora}
\title{Data on Characterization Factors for Eora}
\format{A data frame with 8 columns and 2720 rows}
\usage{
cf_eora
}
\description{
Data on characterization factors for climate change impact, blue water
consumption, energy usage, land use, and material footrpint for EXIOBASE3.
}
\keyword{datasets}
