% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_exio}
\alias{cf_exio}
\title{Data for Characterization Factors on Climate Change}
\format{A data frame with ...}
\usage{
cf_exio
}
\description{
This is data on the characterization factors for climate change.
}
\keyword{datasets}
