% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_exio_multi}
\alias{cf_exio_multi}
\title{Data on Characterization Factors for EXIOBASE}
\format{A list with different elements such as characterization factors for
water stress and characterization components on biodicersity loss.}
\usage{
cf_exio_multi
}
\description{
A compilation of data on characterization factors for water stress, and
biodiversity loss for EXIOBASE3.
}
\keyword{datasets}
