% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Perform basic integrity checks on MRIO Excel file}
\usage{
check(path, precision = 8)
}
\arguments{
\item{path}{Directory of files to check.}

\item{precision}{Largest difference tolerated for equality.}
}
\description{
Given a directory `path`, this will load all Excel files with "MRIO"
(case-insensitive) in the file name and examine them for the following
anomalies:
 \enumerate{
   \item Blank cells.
   \item Negative values for inputs and consumption.
   \item
 }
The number of countries in the MRIO is automatically detected. However, note
that the MRIO table must follow the standard Excel format.
}
