% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_countries.R
\docType{data}
\name{dict_countries}
\alias{dict_countries}
\title{Countries dictionary}
\format{
Data frame with columns
\describe{
\item{iso_num, iso_a2, iso_a3}{Numeric, two-letter, and three-letter country
  codes according to the International Standard for country codes (ISO 3166).
  Retrieved from the ISO Online Browsing Platform as of August 2023.}
\item{un_name}{Country name according to the United Nations.}
\item{adb_code, adb_name}{Country three-letter codes and names according
  to the Asian Development Bank.}
\item{mrio_ind, mrio_ind62, mrio_code, mrio_name}{Country indexing, codes,
  and names according to the ADB Multiregional Input-Output Tables.
  `mrio_ind` is the index for the 72-country version while `mrio_ind62` is
  for the 62-country version.}
\item{wb_name, wb_region}{Country names and regional classifications
  according to the World Bank.}
\item{code, name}{Consolidated country codes and names based on the ADB and
  MRIO conventions and, if missing, the ISO and UN conventions.}
\item{region}{Consolidated regional classification that is essentially the
  World Bank classification with the inclusion of Taipei,China.}
\item{continent}{Country continent; manually assigned.}
}
}
\source{
Asian Development Bank MRIO Team.
}
\usage{
dict_countries
}
\description{
Correspondence table of country names and codes. Note: the use of the term
"country" here is a convenience and intends no judgment on the political
status of the entities listed.
}
\keyword{datasets}
