% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_sectors.R
\docType{data}
\name{dict_sectors}
\alias{dict_sectors}
\title{Sectors dictionary}
\format{
Data frame with columns
\describe{
\item{c_ind, c_name}{Index and name in MRIO.}
\item{c_name_short, c_abv}{More compact alternatives to sector names.}
\item{c15_ind, c15_name}{Correspondence of 35 sectors to 15-sector
aggregation.}
\item{c5_ind, c5_name}{Correspondence of 35 sectors to 5-sector
aggregation.}
}
}
\source{
Asian Development Bank MRIO Team.
}
\usage{
dict_sectors
}
\description{
Correspondence table of sector indices and names and codes in the
Multiregional Input-Output Tables (MRIO).
}
\keyword{datasets}
