% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eurostat.R
\name{download_eurostat}
\alias{download_eurostat}
\title{Download Eurostat macro data}
\usage{
download_eurostat(path, reorder_sheets = TRUE, starting_year = 1995)
}
\arguments{
\item{path}{Path to file or directory. Files must be in Excel and file
names must start with the uppercase three-letter country code. Function
ignores all other files in the directory.}

\item{reorder_sheets}{If \code{TRUE}, arranges the sheets of the workbook
alphabetically upon saving.}

\item{starting_year}{Earliest year of data to download.}
}
\description{
Accesses the Eurostat API using the \code{restatapi} package to
extract macroeconomic data. These are used in the control totals updating
phase of the MRIO compilation process.
}
