% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyads.R
\name{dyads}
\alias{dyads}
\title{Transform matrices into dyadic data.frames}
\usage{
dyads(year, data, country)
}
\arguments{
\item{year}{Numeric vector or number for the respective year/s}

\item{data}{Data object/ matrix that was created before}

\item{input}{String or character vector for input country/ies}

\item{output}{String or character vector for output country/ies}
}
\value{
data.frame
}
\description{
Matrices and list with matrices will be transformed into
data.frames and labeled. The type of data input is recognized
automatically. Year input can be a vector of year or one single year.
Note that high RAM is required to execute this function for all
countries especially for Exiobase and larger time frames.
}
\examples{
dyads(year = 1995:1996, dataobject, output = c("DEU", "CHN"))

}
