\name{eoraloop}
\alias{eoraloop}
\title{Construct list of matrices for specific environmental indicator of Eora}
\usage{
eoraloop(years, indicator)
}
\description{
Load Eora data and return a list of matrices from an environmental indicator ocer a period of time.
}
\examples{
eoraloop(1990:1995, 2)
}
