% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{exioloop}
\alias{exioloop}
\title{Construct list of matrices for specific environmental indicator of EXIOBASE}
\usage{
exioloop(years, indicator)
}
\arguments{
\item{years}{Numeric vector for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator}
}
\value{
Vector of numerics representing the chosen years

Numerics representing the chosen indicator
}
\description{
Load EXIOBASE data and return a list of matrices from an environmental
indicator ocer a period of time.
}
\examples{
exioloop(years = 1995:2000, indicator = 200)

}
