% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get-functions}
\alias{get-functions}
\alias{get_country_name}
\alias{get_country_ind}
\alias{get_country_code}
\alias{get_iso_num}
\alias{get_iso_a2}
\alias{get_iso_a3}
\alias{get_sector_name}
\alias{get_sector_shortname}
\alias{get_sector_ind}
\alias{get_sector_code}
\title{Get MRIO indices, codes, and names}
\usage{
get_country_name(key, unmatched = NA)

get_country_ind(key, unmatched = NA)

get_country_code(key, unmatched = NA)

get_iso_num(key, unmatched = NA)

get_iso_a2(key, unmatched = NA)

get_iso_a3(key, unmatched = NA)

get_sector_name(key, unmatched = NA)

get_sector_shortname(key, unmatched = NA)

get_sector_ind(key, unmatched = NA)

get_sector_code(key, unmatched = NA)
}
\arguments{
\item{key}{An MRIO country or sector index, code, or name. Accepts vectors
and lists.}

\item{unmatched}{What value to set for unmatched keys? Default is \code{NA}.}
}
\value{
An object of the same size as \code{key}.
}
\description{
Functions for easily converting one MRIO key to another.
}
\examples{
get_country_name(41)

# Multiple inputs with different key types (including the target key)
# may be entered
get_country_ind(c("PRC", "Japan", "IND", 62))

# Inputs are not case sensitive
get_country_code("philippines")

}
