% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrio-package.R
\docType{package}
\name{mrio-package}
\alias{mrio}
\alias{mrio-package}
\title{mrio: ADB MRIO Tools}
\description{
Tools for working with the Asian Development Bank (ADB) Multiregional Input-Output (MRIO) tables, a time series of intercountry input-output tables available for up to 72 countries and disaggregated into 35 sectors.
}
\author{
\strong{Maintainer}: Kenneth Anthony Luigi Susmerano Reyes \email{kenneth.rejes@gmail.com}

}
\keyword{internal}
