% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{parse_EB3_SUT}
\alias{parse_EB3_SUT}
\title{Title}
\usage{
parse_EB3_SUT(
  path,
  U = TRUE,
  V = TRUE,
  Y = TRUE,
  va = TRUE,
  metadata = FALSE,
  path2meta = NULL
)
}
\arguments{
\item{metadata}{}
}
\description{
Title
}
