% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_niot.R
\name{pull_niot}
\alias{pull_niot}
\title{Pull national IOT from MRIO}
\usage{
pull_niot(workbook, source, year = NULL)
}
\arguments{
\item{workbook}{Control totals workbook or directory of workbooks. File names
must start with the uppercase three-letter country code. All other files in
the directory are ignored.}

\item{source}{Directory of MRIO tables. Must be stored locally.}

\item{year}{Years of NIOTs to pull. If \code{NULL}, infers years from sheets with
name pattern "5. MRIO 20XX".}
}
\description{
Pulls a national input-output table for a given year from the
Multi-regional Input-Output database and saves it in a control totals
workbook.
}
\details{
Workbook must have at least one sheet with name pattern "5. MRIO
20XX" with the appropriate formatting. Function overwrites any existing
data.
}
