% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{readExio}
\alias{readExio}
\title{Construct matrix for specific environmental indicator of EXIOBASE}
\usage{
readExio(year, indicator)
}
\arguments{
\item{year}{Numeric for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator}

\item{method}{Character string for method to calculate matrix}
}
\value{
Matrix
}
\description{
Load EXIOBASE data and return a matrix from an environmental indicator.
}
\examples{
readExio(year = 1995, indicator = 200,
 method = "demand-production")

}
