% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectorize_comtrade.R
\name{sectorize_comtrade}
\alias{sectorize_comtrade}
\title{Aggregate Comtrade to MRIO sectors}
\usage{
sectorize_comtrade(path)
}
\arguments{
\item{path}{File or directory of files of Comtrade bulk data.}
}
\description{
Aggregates raw product-level data from a UN Comtrade bulk
download to the MRIO sectors.
}
\details{
Each file must be a Comtrade bulk download of one country for one
year compressed via Gzip. The function searches for file names in \code{path}
with the pattern "ABC 20XX.gz", where "ABC" is the uppercase three-letter
country code and "20XX" is the selected year. All other files are ignored.
}
