% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_eora}
\alias{cf_eora}
\title{Data for Characterization Factors for Eora}
\format{A data frame with ... columns and 2720 rows}
\usage{
cf_eora
}
\description{
This is data presents characterization factors for climate change, water
stress, ... for Eora.
}
\keyword{datasets}
