% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_exio}
\alias{cf_exio}
\title{Data on Characterization Factors for EXIOBASE3}
\format{A data frame with 9 columns and 1104 rows}
\usage{
cf_exio
}
\description{
Data on characterization factors for climate change impact, water
stress, blue water consumption, energy usage, land use, and material
footrpint for EXIOBASE3.
}
\keyword{datasets}
