% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf_exio_multi}
\alias{cf_exio_multi}
\title{Data for Characterization Factors for EXIOBASE}
\format{A list with different elements}
\usage{
cf_exio_multi
}
\description{
This is data on the characterization factors for climate change, water
stress, ... for EXIOBASE.
}
\keyword{datasets}
