% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Perform basic integrity checks on MRIO Excel files}
\usage{
check(path, precision = 8)
}
\arguments{
\item{path}{Path to file or directory of files.}

\item{precision}{Number of decimal places to consider in checking whether a
table is balanced.}
}
\description{
Given directory \code{path}, loads all Excel files with "MRIO" (case-insensitive)
in the file name and examines them for anomalies.
}
\details{
The number of countries in the MRIO is automatically detected. However, note
that the MRIO table must follow the standard Excel format.
}
