% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Perform basic integrity checks on MRIO table}
\usage{
check(path, precision = 8, export = TRUE, limit = 10)
}
\arguments{
\item{path}{Path to file or directory. Files must be in Excel and have "MRIO"
(case insensitive) in the name. Function ignores all other files in the
directory.}

\item{precision}{Number of decimal places to consider in checking zero
values.}

\item{export}{If \code{TRUE}, saves results in an Excel file in the same directory
as \code{path}.}

\item{limit}{Number of flagged entries to print. Set to \code{Inf} to print all
entries (not recommended).}
}
\description{
Checks MRIO tables for anomalies. Optionally exports results in
an Excel file. Automatically detects MRIO version (62 vs 72 economies).

List of checks performed:
\enumerate{
\item Non-numeric cells.
\item Blank cells. 1. Negative inputs.
\item Negative consumption items.
\item Negative gross value added (GVA).
\item Negative direct purchases abroad by residents.
\item Positive purchases on the domestic territory by non-residents.
\item Discrepancy between row and column output.
\item Negative sum of value added terms.
\item Negative final sales.
\item Negative total final sales.
\item Negative total exports.
\item GVA = 0 with inputs > 0.
\item GVA = 0 with exports > 0.
}
}
