% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eurostat.R
\name{download_eurostat}
\alias{download_eurostat}
\title{Download Eurostat macro data}
\usage{
download_eurostat(workbook, reorder_sheets = TRUE, starting_year = 1995)
}
\arguments{
\item{workbook}{Control totals workbook or directory of workbooks. File names
must start with the uppercase three-letter country code. All other files in
the directory are ignored.}

\item{reorder_sheets}{If \code{TRUE}, arranges the sheets of the workbook
alphabetically upon saving.}

\item{starting_year}{Earliest year of data to download.}
}
\description{
Downloads macroeconomic data from Eurostat and saves them in a
control totals workbook.
}
\details{
Uses the \code{restatapi} package to access the Eurostat API. Downloaded
data are saved in a series of sheets with name pattern "1. ES \if{html}{\out{<dataset
code>}}". Sheets are created if they do not exist.
}
