% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyads.R
\name{dyads}
\alias{dyads}
\title{Transform matrices into dyadic data.frames}
\usage{
dyads(year, data)
}
\arguments{
\item{year}{Numeric vector or number for the respective year/s}

\item{data}{Data object/ matrix that was created before}
}
\value{
None
}
\description{
Matrices and list objects with EEMRIO matrices will be transformed into
data.frames and labeled. The type of data input is recognized
automatically. Year input can be a vector of year or one single year.
Note that high RAM is required to execute this function for Exiobase and
multiple years.
}
\examples{
dyads(year = 1995:1996, eoramatrixobject)

}
