% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eora.R
\name{eoraloop}
\alias{eoraloop}
\title{Construct list of matrices for specific environmental indicator of Eora}
\usage{
eoraloop(years, indicator)
}
\arguments{
\item{years}{Numeric vector for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator}

\item{method}{Character string for method to calculate matrix}
}
\value{
Matrix or list
}
\description{
Load Eora data and return a list of matrices from an environmental
indicator ocer a period of time.
}
\examples{
eoraloop(years = 1995:2000, indicator = 200)

}
