% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eora.R
\name{eoraloop}
\alias{eoraloop}
\title{Construct list of matrices for specific environmental indicator of Eora}
\usage{
eoraloop(years, indicator)
}
\arguments{
\item{years}{Numeric vector for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator or character string for characterization factor
\itemize{
\item "bl" for \strong{b}iodiversity \strong{l}oss
\item "bw" for \strong{b}lue \strong{w}ater consumption
\item "cc" for \strong{c}limate \strong{c}hange impacts
\item "en" for \strong{en}ergy demand
\item "lu" for \strong{l}and \strong{u}se
\item "mf" for \strong{m}aterial \strong{f}ootprint
\item "ws" for \strong{w}ater \strong{s}tress
}}

\item{method}{Character string for method to calculate matrix}
}
\value{
Matrix or list
}
\description{
Load Eora data and return a list of matrices from an environmental
indicator ocer a period of time.
}
\examples{
eoraloop(years = 1995:2000, indicator = 200)

}
