% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exiobase.R
\name{exioloop}
\alias{exioloop}
\title{Construct list of matrices for specific environmental indicator of EXIOBASE3}
\usage{
exioloop(years, indicator, method, target)
}
\arguments{
\item{years}{Numeric vector for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator or character string for characterization factor
\itemize{
\item "bl" for \strong{b}iodiversity \strong{l}oss
\item "bw" for \strong{b}lue \strong{w}ater consumption
\item "cc" for \strong{c}limate \strong{c}hange impacts
\item "en" for \strong{en}ergy demand
\item "lu" for \strong{l}and \strong{u}se
\item "mf" for \strong{m}aterial \strong{f}ootprint
\item "ws" for \strong{w}ater \strong{s}tress
}}

\item{method}{Character string for method to calculate matrix
\itemize{
\item "pd" for \strong{p}roduction to \strong{d}emand matrix
\item "no-double-pt" for \strong{p}roduction to target \strong{d}emand matrix
\item "no-double-ts" for \strong{t}arget to final \strong{s}upply matrix
\item "no-double-td" for \strong{t}arget to final \strong{d}emand matrix
\item "no-double-pd" for \strong{p}roduction to final \strong{d}emand matrix
}}

\item{target}{Three letter country code for corresponding target
country if double counting is prevented}

\item{type}{Which type of matrix should be used for calculation: industry to
industry ("ixi") or product to product ("pxp"). Default is type = "ixi".
Make sure you have downloaded pxp-matrices if you want to work with them.}
}
\value{
Matrix or list
}
\description{
Load EXIOBASE3 data and return a list of matrices from an environmental
indicator over a period of time.
}
\examples{
exioloop(years = 1995:2000, indicator = 200)
exioloop(year = 1995:1997, indicator = "cc", method = "pd",
target = "USA")

}
