\name{exioloop}
\alias{exioloop}
\title{Construct list of matrices for specific environmental indicator of EXIOBASE}
\usage{
exioloop(years, indicator)
}
\description{
Load EXIOBASE data and return a list of matrices from an environmental indicator ocer a period of time.}
\examples{
exioloop(1995:2000, 2)
}

