% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eora.R
\name{readEora}
\alias{readEora}
\title{Construct matrix for specific environmental indicator of Eora}
\usage{
readEora(year, indicator)
}
\arguments{
\item{year}{Numeric for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator or character string for characterization factor
\itemize{
\item "bl" for \strong{b}iodiversity \strong{l}oss
\item "bw" for \strong{b}lue \strong{w}ater consumption
\item "cc" for \strong{c}limate \strong{c}hange impacts
\item "en" for \strong{en}ergy demand
\item "lu" for \strong{l}and \strong{u}se
\item "mf" for \strong{m}aterial \strong{f}ootprint
\item "ws" for \strong{w}ater \strong{s}tress
}}

\item{method}{Character string for method to calculate matrix}
}
\value{
Matrix
}
\description{
Load Eora data and return a matrix from an environmental indicator.
}
\examples{
readEora(year = 1995, indicator = 200)

}
