% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eora.R
\name{readEora}
\alias{readEora}
\title{Construct matrix for specific environmental indicator of Eora}
\usage{
readEora(year, indicator)
}
\arguments{
\item{year}{Numeric for the respective year}

\item{indicator}{Numeric for the row number of the corresponding
indicator or character string for characterization factor
* "cc" for **c**limate **c**hange impacts
* "ws" for **w**ater **s**tress
* "lu" for **l**and **u**se}

\item{method}{Character string for method to calculate matrix}
}
\value{
Matrix
}
\description{
Load Eora data and return a matrix from an environmental indicator.
}
\examples{
readEora(year = 1995, indicator = 200)

}
